+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+-   Add "Next Fruit" label above the next fruit in HUD
+
 ## [0.1.0] - 2023-11-30
 
 ### Changed